Using turbo frame to prevent form's reloading behavior on submitting, using turbo stream to append created exercise card to the exercise list. Added form reset feature using stimulus. Move tom select stimulus code from the seperated controller to the general form js controller to integrate tom select #clear method into the reset method.
