+++ app/controllers/session_exercises_controller.rb
@@ -6,7 +6,13 @@ class SessionExercisesController < ApplicationController
     @session_exercise.workout_session = @workout_session
     authorize @session_exercise
     if @session_exercise.save
-      redirect_to workout_session_path(@workout_session), notice: "Exercise added successfully"
+      respond_to do |format|
+        format.turbo_stream do
+          render turbo_stream: turbo_stream.append(:session_exercises, partial: "workout_sessions/exercise_display",
+                                                                       locals: { session_exercise: @session_exercise })
+        end
+        format.html { redirect_to workout_session_path(@workout_session), notice: "Exercise added successfully" }
+      end
     else
       @session_exercises = @workout_session.session_exercises if @workout_session.session_exercises.any?
       @comment = Comment.new
