+++ app/views/workout_sessions/_new_exercise.html.erb
@@ -1,77 +1,76 @@
-<div class="container py-5">
-  <div class="row justify-content-center">
-    <div class="col-lg-10 col-md-10 col-sm-12">
-      <div class="card border-0 shadow-sm">
-        <div class="card-header bg-light border-0 py-4">
-          <h1 class="h3 text-center mb-2 fw-bold">Add exercises</h1>
-          <p class="text-muted text-center mb-0">Customize your client's exercise</p>
-        </div>
-        <div class="card-body p-5">
-          <%= simple_form_for([workout_session, session_exercise], data: { controller: "form-prevent-default", action: "ajax:afterSend->form-prevent-default#submit" }) do |f| %>
-            <div class="row g-4">
-              <!-- Exercise Selection -->
-              <!-- For using Tom Select -->
-              <!-- 1. importmap pin tom-select -> go to importmap.rb, change the version source if needed -->
-              <!-- 2. generate stimulus controller and mount the controller on the select input field-->
-              <!-- 3. add a line to import tom select into the stimulus controller -->
-              <!-- 4. In application.html.erb, add a link for tom select stylesheet-->
-              <div class="col-12">
-                <%= f.input :exercise_id,
+<%= turbo_frame_tag :new_exercise_form do %>
+  <div class="container py-5">
+    <div class="row justify-content-center">
+      <div class="col-lg-10 col-md-10 col-sm-12">
+        <div class="card border-0 shadow-sm">
+          <div class="card-header bg-light border-0 py-4">
+            <h1 class="h3 text-center mb-2 fw-bold">Add exercises</h1>
+            <p class="text-muted text-center mb-0">Customize your client's exercise</p>
+          </div>
+          <div class="card-body p-5">
+            <%= simple_form_for([workout_session, session_exercise], data: { controller: "session-exercise-form", action: "turbo:submit-end->session-exercise-form#reset" }) do |f| %>
+              <div class="row g-4">
+                <!-- Exercise Selection -->
+                <!-- For using Tom Select -->
+                <!-- 1. importmap pin tom-select -> go to importmap.rb, change the version source if needed -->
+                <!-- 2. generate stimulus controller and mount the controller on the select input field-->
+                <!-- 3. add a line to import tom select into the stimulus controller -->
+                <!-- 4. In application.html.erb, add a link for tom select stylesheet-->
+                <div class="col-12">
+                  <%= f.input :exercise_id,
                     collection: Exercise.all.map { |exercise| [ "#{exercise.name}", exercise.id ] },
                     label: "Which exercise are you adding?",
                     input_html: {
                       id: "select-exercise",
-                      data: { controller: "tom-select", target: "inputField" },
+                      data: { session_exercise_form_target: "tomSelect"},
                       class: "form-select form-select-lg"
                     } %>
-              </div>
-              <!-- Load -->
-              <div class="col-4">
-                <div class="input-group mb-3">
-                  <%= f.input :load,
+                </div>
+                <!-- Load -->
+                <div class="col-4">
+                  <div class="input-group mb-3">
+                    <%= f.input :load,
                       label: "Load:",
                       input_html: {
                         class: "form-control",
-                        data: { target: "inputField" },
                         placeholder: "10"
                       }, wrapper: false %>
-                  <span class="input-group-text">kg</span>
+                    <span class="input-group-text">kg</span>
+                  </div>
                 </div>
-              </div>
-              <!-- Repetitions -->
-              <div class="col-4">
-                <div class="input-group mb-3">
-                  <%= f.input :repetitions,
+                <!-- Repetitions -->
+                <div class="col-4">
+                  <div class="input-group mb-3">
+                    <%= f.input :repetitions,
                       label: "Reps:",
                       input_html: {
                         class: "form-control",
-                        data: { target: "inputField" },
                         placeholder: "8"
                       }, wrapper: false %>
-                  <span class="input-group-text">reps</span>
+                    <span class="input-group-text">reps</span>
+                  </div>
                 </div>
-              </div>
-              <!-- Sets -->
-              <div class="col-4">
-                <div class="input-group mb-3">
-                  <%= f.input :set,
+                <!-- Sets -->
+                <div class="col-4">
+                  <div class="input-group mb-3">
+                    <%= f.input :set,
                       label: "Sets:",
                       input_html: {
                         class: "form-control",
-                        data: { target: "inputField" },
                         placeholder: "3"
                       }, wrapper: false %>
-                  <span class="input-group-text">sets</span>
+                    <span class="input-group-text">sets</span>
+                  </div>
+                </div>
+                <!-- Submit Button -->
+                <div class="col-12 text-center pt-3">
+                  <%= f.button :submit, "Add this exercise", class: "btn btn-primary btn-lg px-5" %>
                 </div>
               </div>
-              <!-- Submit Button -->
-              <div class="col-12 text-center pt-3">
-                <%= f.button :submit, "Add this exercise", class: "btn btn-primary btn-lg px-5" %>
-              </div>
-            </div>
-          <% end %>
+            <% end %>
+          </div>
         </div>
       </div>
     </div>
   </div>
-</div>
+<% end %>
